Silly capistrano upcases your environment variables
So our default_env fix to add the SERVER_URL in capistrano didn't work.
Have to set it to upper case and user an upper case variable in puppet
(weird).
Also added a config line to avoid loading the DB for assets pre-compile
that was somehow complaining about sqlite3 in the production environment
setting.