Close #141
This was a miss when upgrading devise a while back.
Passwords could not be correctly reset before this change because devise
now saves an encoded version of the password in the DB and hands over
the original token. Therefore, the token sent in the email cannot be the
one in the DB.