Styling flash boxes per class and ensuring ids are unique
Rails now returns an array of errors for errors[:key] since one can have
more than one error in a single key. So a flash error can contain an
array of values instead of a single value. Therefore, we iterate over
the contents of a flash value in case it is an array and generate a div
per content with an incremental ID and a class to identify the groups.