No longer use seed data in tests.
As we start having more than one conference using the system, the seed
data will change and relying on it for tests to pass is not recommended.
This closes issue #155.

It also closes issue #156 as changes were done to replace the old rspec
assertion syntax from object.should to expect(object).to. Also used the
occasion to upgrade the hash syntax from ':symbol =>' to 'symbol:' for Ruby
1.9+ compatibility only.
Changes were made with regular expression replacement:
/( *)(.*)\.should(.*)/ replaced with '\1expect(\2).to\3'
/\.to == (.*)/ replaced with '.to eq(\1)'
/:(\w+) =>/ replaced with '\1:'