Fixing user registration
Devise is compatible with rails 4 strong parameters. It detects if
strong parameters are available and, if so, it uses that to filter which
parameters should be allowed to go through the registration and sign in
pages. As I added strong parameters (the gem) for issue #101, devise
stopped accepting our custom parameters for user registrations.
This allows them to be set.