Close #153.
Json extension is the default so no need to expect it.
Removed the super class API Controller as it didn't make sense for the
sessions.
Migrated the gravatar_url to the application controller since it was
needed for the JSON.
Show only for now. Will probably want an API for accepted sessions for a
given conference at some point. And maybe an RSS feed URL to listen for
changes on those sessions/conference. Will have to create another issue
for that.