Still working on reviewer invitation process
It is now easier to invite people that were reviewers in previous
conferences. This is mostly how organizers will be able to check on
feedback given to reviewers (for issue #159) so that they can get a
sense of how people are doing.
Next steps are:
1) Improve this new page as well as the reviewer 'My reviews' page to
provide an overview of the feedback received.
2) Create the reviewer show page that will have detailed information on
the feedback as well as their reviews. Link that from the summarized
previous view.