Closes issue #161
Quite some changes in the way to view your own reviews but pretty much
includes the same information plus the feedback.
Also re-used the idea of presenting a summary of all session data in the
organizer feedback visualization to better use space.
Also did a small performance improvement in the reviews visualization
page.