Adding top commenters public API
Closes #170 by adding an API accessible at
/api/top_commenters.json?limit=10&filter%5Byear%5D=2015
It defaults to 5 top commenters of all times.
Limit can be passed as a query parameter with any value between 1 and
20.
A year filter is available through a query param filter[year] and should
correspond to a conference year (2010 to 2015). It'll count comments
only for sessions from that conference.
Results are a JSON representation of a user with username, full name,
gravatar_url and number of comments. Ideally this would include links to
the profile but this first version is good enough.