Fixing the reviewer agreement terms acceptance
Turns out that as part of #114, `validates :field, acceptance: true`
works great for fields that are NOT under ActiveRecord's control (such
as reviewer_agreement since you can only be a reviewer if you agree).
However, it requires a change for fields that ARE under ActiveRecord's
control (such as author_agreement in a Session).
As part of 8ab435d2, I found the problem with Session and fixed it.
However, since the construct was the same for Reviewer, I made the
mistake of changing it to `validates :reviewer_agreement, acceptance: {
accept: true}` too. Unfortunately, values coming from the browser are
left as '0' and '1' instead of false or true. As a result, browser
submissions of accepting the reviewer agreement would not validate.