Using non-hardcoded info for session types requirements
Now we use data from DB to decide which session types should react by
showing or hiding audience_limit and which ones should require (or not)
mechanics.
This completes #223 for the future.
There is still a pending refactoring to remove hard-coded tracks that
require specific session_types. Since this feature has not been used
after 2012, this might just be a legacy code that will be killed
eventually once DB conference fixtures are removed.