Ensure the tagcloud is visible when rendering.
This allows the body to be like:
```
div(tagcloud){width: 700px;height: 300px;display:none;}. &nbsp;
```
If the tagcloud doesn't have anything to show, the body of the page
stays in the same spot. If it has something, it renders like an image
would by pushing the content down when available.